Executor keepalive thread

+++ deployer/src/main/java/org/jboss/fuse/qa/fafram8/cluster/container/SshContainer.java
@@ -6,6 +6,8 @@ import org.jboss.fuse.qa.fafram8.deployer.ContainerSummoner;
 import org.jboss.fuse.qa.fafram8.exception.FaframException;
 import org.jboss.fuse.qa.fafram8.executor.Executor;
 import org.jboss.fuse.qa.fafram8.manager.ContainerManager;
+import org.jboss.fuse.qa.fafram8.modifier.ModifierExecutor;
+import org.jboss.fuse.qa.fafram8.property.FaframConstant;
 import org.jboss.fuse.qa.fafram8.property.SystemProperty;
 import org.jboss.fuse.qa.fafram8.provision.provider.ProviderSingleton;
 import org.jboss.fuse.qa.fafram8.util.Option;
@@ -74,7 +76,7 @@ public class SshContainer extends Container implements ThreadContainer {
 		}
 
 		// If using static provider then clean
-		if (ProviderSingleton.INSTANCE.isStaticProvider()) {
+		if (ProviderSingleton.INSTANCE.isStaticProvider() && !SystemProperty.isWithoutPublicIp()) {
 			clean();
 		}
 
@@ -92,10 +94,20 @@ public class SshContainer extends Container implements ThreadContainer {
 		}
 
 		log.info("Creating container " + this);
-		executor.connect();
-		// Recreate the executor because the values could be changed in the process
-		super.getNode().setExecutor(super.getNode().createExecutor());
+		if (!executor.isConnected()) {
+			log.trace("Connecting executor " + executor + " before creating ssh container");
+			executor.connect();
+		}
 
+		if (!SystemProperty.isWithoutPublicIp()) {
+			// Connect the node executor before executing the container-create-ssh command - that will ensure that the machine is up
+			if (!super.getNode().getExecutor().isConnected()) {
+				log.trace("First time connecting node executor");
+				super.getNode().getExecutor().connect();
+			}
+		} else {
+			log.warn(FaframConstant.WITHOUT_PUBLIC_IP + " is set, node won't be available");
+		}
 		executor.executeCommand(String.format("container-create-ssh --user %s --password %s --host %s %s %s",
 				super.getNode().getUsername(), super.getNode().getPassword(), super.getNode().getHost(), OptionUtils.getCommand(super.getOptions()), super.getName()));
 		super.setCreated(true);
@@ -106,14 +118,18 @@ public class SshContainer extends Container implements ThreadContainer {
 			throw e;
 		}
 
-		super.setExecutor(super.createExecutor());
-		super.getExecutor().connect();
-		super.getNode().getExecutor().connect();
+		if (!SystemProperty.isWithoutPublicIp()) {
+			log.trace("First time connecting ssh executor");
+			super.setExecutor(super.createExecutor());
+			super.getExecutor().connect();
+		} else {
+			log.warn(FaframConstant.WITHOUT_PUBLIC_IP + " is set, executeCommand / executeNodeCommand won't work");
+		}
 		super.setOnline(true);
 		// Set the fuse path
 		try {
 			super.setFusePath(super.getExecutor().executeCommandSilently("shell:info | grep \"Karaf base\"").trim().replaceAll(" +", " ")
-					.split(" ")[1]);
+					.split(" ")[2]);
 		} catch (Exception ex) {
 			log.warn("Setting fuse path failed, it won't be available", ex);
 		}
@@ -130,11 +146,25 @@ public class SshContainer extends Container implements ThreadContainer {
 			return;
 		}
 
+		if (!SystemProperty.isWithoutPublicIp()) {
+			ModifierExecutor.executePostModifiers(this, super.getNode().getExecutor());
+
+			super.getNode().getExecutor().stopKeepAliveTimer();
+			super.getExecutor().stopKeepAliveTimer();
+		}
 		log.info("Destroying container " + super.getName());
-		executor.connect();
-		super.getParent().getExecutor().executeCommand("container-delete --force " + super.getName());
+		if (!executor.isConnected()) {
+			log.trace("Connecting executor " + executor + " before deleting ssh container");
+			executor.connect();
+		}
+		executor.executeCommand("container-delete --force " + super.getName());
 		super.setCreated(false);
 		ContainerManager.getContainerList().remove(this);
+		if (!SystemProperty.isWithoutPublicIp()) {
+			log.trace("Disconnecting node/fuse executors after destroying the container");
+			super.getExecutor().disconnect();
+			super.getNode().getExecutor().disconnect();
+		}
 	}
 
 	@Override
@@ -148,6 +178,10 @@ public class SshContainer extends Container implements ThreadContainer {
 		super.getParent().getExecutor().executeCommand("container-start " + (force ? "--force " : "") + super.getName());
 		super.getParent().getExecutor().waitForProvisioning(this);
 		super.setOnline(true);
+		if (!SystemProperty.isWithoutPublicIp()) {
+			log.trace("Connecting executor in ssh's start()");
+			super.getExecutor().connect();
+		}
 	}
 
 	@Override
@@ -155,12 +189,22 @@ public class SshContainer extends Container implements ThreadContainer {
 		super.getParent().getExecutor().executeCommand("container-stop " + (force ? "--force " : "") + super.getName());
 		super.getParent().getExecutor().waitForContainerStop(this);
 		super.setOnline(false);
+		if (!SystemProperty.isWithoutPublicIp()) {
+			log.trace("Disconnecting executor in ssh's stop()");
+			super.getExecutor().disconnect();
+		}
 	}
 
 	@Override
 	public void kill() {
-		super.getNode().getExecutor().executeCommand("pkill -9 -f " + super.getName());
-		super.setOnline(false);
+		if (!SystemProperty.isWithoutPublicIp()) {
+			super.getNode().getExecutor().executeCommand("pkill -9 -f " + super.getName());
+			super.setOnline(false);
+			log.trace("Disconnecting executor in ssh's kill()");
+			super.getExecutor().disconnect();
+		} else {
+			log.warn(FaframConstant.WITHOUT_PUBLIC_IP + " is set, kill won't work");
+		}
 	}
 
 	@Override
@@ -211,12 +255,13 @@ public class SshContainer extends Container implements ThreadContainer {
 		final String path;
 		final String workDir = OptionUtils.getString(super.getOptions(), Option.WORKING_DIRECTORY);
 		if (!("".equals(workDir)) || !("".equals(SystemProperty.getWorkingDirectory()))) {
-//			 Decide if working directory was set on ssh and if not set the system property as default
+			// Decide if working directory was set on ssh and if not set the system property as default
 			path = "".equals(workDir) ? SystemProperty.getWorkingDirectory() : workDir;
 		} else {
 			path = "containers";
 		}
-//		 Executor needs to be connected before executing command
+		// Executor needs to be connected before executing command
+		log.trace("Connecting ssh node executor before cleaning the node");
 		super.getNode().getExecutor().connect();
 		super.getNode().getExecutor().executeCommand("rm -rf " + path + File.separator + super.getName());
 	}
